+++ Changelog.md
@@ -1,5 +1,7 @@
 ## 0.0.26
-*INCOMPLETE*
++ Refactored `fxn.predictions.to_value` method to `to_object` for clarity.
++ Refactored `fxn.predictions.from_value` method to `to_value` for clarity.
++ Updated `fxn.predictions.to_object` method to always use aliased field names when serializing Pydantic types.
 
 ## 0.0.25
 + Fixed JSON serialization errors when using the CLI to perform some operations.